backend/requirements.txt
fastapi
uvicorn[standard]
httpx
psycopg2-binary
alembic
SQLAlchemy
python-dotenv
APScheduler
pydantic